Why Ready to Move Flats Are So Demanded in Gurgaon
1. Introduction –

When it comes to the state of affairs in Gurgaon, Gurgaon real estate is one of the most popular and sought-after asset classes in India. With the continuous upward trend of the industry, it’s no surprise that real estate has been considered as the most lucrative investment asset class.
It is no surprise that investment in real estate has become a very attractive option for the investors and this is evident from the fact that there are several projects in Gurgaon being developed buildmakaan.in.
Bilderfloor is one such project which was set up by an American company BilderFloor LLC with a view toward providing a professional service to residents of Gurgaon. This innovative development concept offers affordable housing options ranging from single family homes to studios and townhouses on an attractive plot of land available in Gurgaon. The project also offers residential facilities like swimming pools, gymnasiums, meeting rooms etc along with commercial amenities like restaurants, supermarkets etc.
The project can be conveniently reached by road or rail and there are various railway stations located within walking distance of this project. Besides this, some other major public transport facilities including STC airport bus terminal and metro station are also nearby which makes it very convenient for all kinds of travelers who live in this area.
With such a wide range of amenities at its disposal; Bilderfloor LLC has become one of the leaders in offering affordable housing options at a relatively low cost to its customers.

2. Thesis Statement

Ready to Move flats are the most wanted property in Gurgaon. It is a city where every company has a big presence as a result of its business activities. But the lack of housing options has made it difficult for them to find accommodation. In fact, the demand for such flats is not limited to just Gurgaon alone, but also the entire country.
The concept of 'ready-to-move' flats was first introduced by Hinduja group which boasts one of the largest investments in India as they have been investing in real estate and aviation sector since last 15 years and after their entry into Gurgaon market, they have successfully floated ready-to-move (RWM) project named 'Majestic' building with an area of 7 lakh sq ft. The complete project has been developed by TCS and it is expected that this will be one of the largest housing projects in Gurgaon to provide affordable housing option for all sections of society. Login buildmakaan.in

3. Body Paragraphs

So, the question is, why is Gurgaon so popular for the real estate industry and why does it attract so many buyers?
Surprisingly, Gurgaon has a lot of land where you can build your own house. It’s not just a place to live. There are a lot of large residential buildings there. You can buy land and build your own house like in any other country.
But most people don’t know that you can buy ready-to-move flats in Gurgaon as well. Buildmakaan.com As it is an affordable place these flats are very popular among the investors since they save a lot of money and can easily afford them as compared to buying an expensive home in Delhi or Mumbai.
You need to be aware about the rapid growth of housing demand in Gurgaon for different reasons which explain how so many people are looking for homes there that are ready-to-move but at the same time, not quite ready to move into them. So what will happen after one move? When will it be ready to move and when will it be ready to stay?
As per the population of Gurgaon, it has been predicted that by 2025, there will be 1 million more people coming from outside whereas at present Delhi has around 1 million more people than Gurgaon alone and Mumbai has around 1 million more than both Delhi and Mumbai combined. That gives rise to the question “When is it going to be ready?”
There are various factors which contribute to this increase in housing demand including the economic growth, growth in employment opportunities with industries located there like automobile manufacturing industry which offers employment opportunities for young engineers etc., higher disposable income available due to higher income tax bracket as well as immigration from other countries for different purposes like education etc., increasing family size due to growing population etc., increasing number of houses being built every year by different builders etc.. And on top of all this higher demand comes from higher price rates as land prices go up with construction going on faster than before.buildmakaan.in  
4. Conclusion

Real estate is a profitable industry and considered as one of the fastest growing markets all over the world. No wonder, in Gurgaon real estate has become one of the most sought-after assets for investors.day by day new home build in gurgaon with buildmakaan.com
The most important thing is to be aware of the rules and regulations governing real estate in Gurgaon before investing in this sector. The city can be divided into three distinct zones:
Gargi Zone – This zone is the older part of Gurgaon and offers high-end residential spaces with good infrastructure like roads, residential complexes, hospitals, schools etc. It also has good commercial areas and numerous hospitals.
Vishakhpur Zone – This zone covers the area where industrial activities are taking place and has many commercial buildings that cater to various industries such as power generation, cement plants etc. The economical aspect of this zone is also very good as it offers its residents cheap electricity bills at a low price.
Gurgaon Industrial Area – This zone covers industrial areas like Aligarh Industrial Area, Damansara Industrial Area etc., which are mostly owned by international companies. They are located in this zone close to residential areas and offer you cheap electricity bills at a low price.
In addition to these three zones there is another one called DND Zone which provides access to localities such as Juhu Vihar, Daryaganj etc., which gives you better options when purchasing a flat from these localities. However, if you look at these areas from an investment perspective then it becomes clear that they do not make any sense for someone who is looking for a flat in Gurgaon only buildmakaan.com
The other point worth mentioning is that there are many more areas that can be categorized under Gurgaon but when it comes to buying property in Gurgaon then it’s best to avoid those places where there’s no activity or where there’s no infrastructure or experts working on building projects etc., so that your investment doesn’t get affected due to lack of security or other factors.